Move test exclusion to common library.
