+++ Code/Tools/TestImpactFramework/Runtime/Native/Code/Include/TestImpactFramework/Native/TestImpactNativeConfiguration.h
@@ -8,7 +8,6 @@
 
 #pragma once
 
-#include <TestImpactFramework/TestImpactTestSequence.h>
 #include <TestImpactFramework/TestImpactConfiguration.h>
 
 namespace TestImpact
